Backend/requirements.txt
absl-py==0.11.0
Django==3.0.3
django-ratings==0.3.7
djangorestframework==3.11.0
gunicorn==20.0.4
pip-tools==5.5.0
pip-upgrade-tool==0.4.5
pipenv==2020.11.15
portalocker==2.1.0
sqlparse==0.4.1
whitenoise==5.2.0
